Replace composition with inheritance
Pull AbstractSourceBuilder out of SourceStringBuilder and use it to remove composition from CompilationUnitBuilder. This seems counter to good practice, but the aim is to merge the implementations into a single type.
